Create an RSS feed.
Required adding a short description to every post. There were actual tabs in
some of the posts, which broke the frontmatter parser, so I replaced all tabs
with spaces.

Also updated all links to explicitly use HTTPS (if the remote site supports it),
and fixing a few other small issues in posts.